fixed permissions

+++ src/main/java/org/example/crm/entity/analyticsRecord/AnalyticBranch.java
@@ -1,6 +1,6 @@
 package org.example.crm.entity.analyticsRecord;
 
 public record AnalyticBranch(
-        Integer branchCount
+        Long branchCount
 ) {
 }
