+++ src/main/java/org/example/crm/entity/analyticsRecord/AnalyticStudent.java
@@ -1,9 +1,9 @@
 package org.example.crm.entity.analyticsRecord;
 
 public record AnalyticStudent(
-        Integer studentCount,
-        Integer studentsAddedInPrevMonth,
-        Integer studentsAddedInMonth,
+        Long studentCount,
+        Long studentsAddedInPrevMonth,
+        Long studentsAddedInMonth,
         double difference
 ) {
 }
