+++ src/main/java/org/example/crm/projection/AnalyticBranchProjection.java
@@ -1,5 +1,5 @@
 package org.example.crm.projection;
 
 public interface AnalyticBranchProjection {
-    Integer getBranchCount();
+    Long getBranchCount();
 }
