+++ src/main/java/org/example/crm/projection/AnalyticEnrollmentProjection.java
@@ -1,7 +1,7 @@
 package org.example.crm.projection;
 
 public interface AnalyticEnrollmentProjection {
-    Integer getEnrollmentCount();
-    Integer getEnrollmentCountInMonth();
-    Integer getEnrollmentCountInPreviousMonth();
+    Long getEnrollmentCount();
+    Long getEnrollmentCountInMonth();
+    Long getEnrollmentCountInPreviousMonth();
 }
