+++ src/main/java/org/example/crm/projection/AnalyticLeadProjection.java
@@ -1,7 +1,7 @@
 package org.example.crm.projection;
 
 public interface AnalyticLeadProjection {
-    Integer getLeadCount();
-    Integer getLeadCountInMonth();
-    Integer getLeadCountInPrevMonth();
+    Long getLeadCount();
+    Long getLeadCountInMonth();
+    Long getLeadCountInPrevMonth();
 }
