+++ src/main/java/org/example/crm/projection/AnalyticStudentProjection.java
@@ -1,7 +1,7 @@
 package org.example.crm.projection;
 
 public interface AnalyticStudentProjection {
-    Integer getStudentCount();
-    Integer getStudentsAddedInMonth();
-    Integer getStudentsAddedInPrevMonth();
+    Long getStudentCount();
+    Long getStudentsAddedInMonth();
+    Long getStudentsAddedInPrevMonth();
 }
