+++ src/main/java/org/example/crm/projection/AnalyticTeacherProjection.java
@@ -1,7 +1,7 @@
 package org.example.crm.projection;
 
 public interface AnalyticTeacherProjection {
-    Integer getTeacherCount();
-    Integer getTeachersAddedInMonth();
-    Integer getTeachersAddedInPrevMonth();
+    Long getTeacherCount();
+    Long getTeachersAddedInMonth();
+    Long getTeachersAddedInPrevMonth();
 }
